dags/pyproject.toml
[tool.poetry]
name = "seedflow"
version = "0.1.0"
description = ""
authors = ["ChethanUK <chethanuk@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
airflow-provider-great-expectations = "^0.2.4"
airflow-provider-duckdb = "^0.0.1"
astro-sdk-python = {extras = ["postgres"], version = "^1.4.1"}

[tool.poetry.dev-dependencies]
apache-airflow = {extras = ["postgres", "virtualenv"], version = "^2.5.1"}
pytest = "^7.2.1"
duckdb = "^0.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
